+++ lib/distribute_reads.rb
@@ -18,14 +18,14 @@ module DistributeReads
     lag_failover: false
   }
 
-  def self.lag
-    conn = ActiveRecord::Base.connection
-    if %w(PostgreSQL PostGIS).include?(conn.adapter_name)
-      if conn.instance_variable_get(:@slave_pool).connections.size > 1
+  def self.lag(connection: nil)
+    connection ||= ActiveRecord::Base.connection
+    if %w(PostgreSQL PostGIS).include?(connection.adapter_name)
+      if connection.instance_variable_get(:@slave_pool).connections.size > 1
         warn "[distribute_reads] Multiple replicas available, lag only reported for one"
       end
 
-      conn.execute(
+      connection.execute(
         "SELECT CASE
           WHEN NOT pg_is_in_recovery() OR pg_last_xlog_receive_location() = pg_last_xlog_replay_location() THEN 0
           ELSE EXTRACT (EPOCH FROM NOW() - pg_last_xact_replay_timestamp())